requirements.txt
aiogram==2.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
babel==2.16.0
black==24.10.0
certifi==2024.12.14
click==8.1.8
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
pytest==8.3.4
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
